refactor(inspect): split InspectorOverlay paint methods into a sibling TU (P11-5, #2647)
inspector_overlay.cpp was 2,102 lines — over the long-tail 2,000-line target.
The paint cluster (paint + 12 paint_* methods + the paint-only tweak_action_at
helper + the preview_value/abbreviate_anchor anon-namespace, 15 method bodies /
~1,260 lines) is the dominant concern; carve it verbatim into
inspector_overlay_paint.cpp, following the #2555 internal-header pattern.

The one shared file-local static — color_to_hex (used by both the eyedropper
pick path that stays and the paint TU) — moves to inspector_overlay_internal.hpp
as an `inline` helper.

inspector_overlay.cpp: 2,102 -> 836 lines. Built GPU-enabled (arm64, pinned
Skia/Dawn); pulp-test-inspector passes 277 assertions / 54 cases, unchanged.

Closes the P11-5 long-tail sweep (test_inspector #2666, test_host #2677, and
now inspector_overlay all under 2,000 lines).
