+++ inspect/CMakeLists.txt
@@ -12,6 +12,7 @@ target_sources(pulp-inspect PRIVATE
     src/inspector_overlay_field_edit.cpp
     src/inspector_overlay_zoom.cpp
     src/inspector_overlay_pass_viewer.cpp
+    src/inspector_overlay_paint.cpp
     src/inspector_window.cpp
     src/state_inspector.cpp
     src/console_capture.cpp
