+++ inspect/src/inspector_overlay_paint.cpp
@@ -1,0 +1,1281 @@
+// InspectorOverlay paint methods, split out of inspector_overlay.cpp (P11-5,
+// #2647) to bring the 2,102-line parent under the 2,000-line target. Follows
+// the #2555 internal-header pattern: shared palette + color_to_hex live in
+// inspector_overlay_internal.hpp. Includes the parent's own helper anon-namespace
+// (preview_value/abbreviate_anchor) used only by these paint routines.
+#include "inspector_overlay_internal.hpp"
+
+#include <pulp/inspect/inspector_overlay.hpp>
+#include <pulp/inspect/tweak_store.hpp>
+#include <pulp/view/inspector.hpp>
+#include <pulp/render/render_pass.hpp>
+#include <pulp/render/atlas_inventory.hpp>
+
+#include <choc/text/choc_JSON.h>
+
+#include <algorithm>
+#include <cstdlib>
+#include <sstream>
+#include <iomanip>
+#include <cmath>
+
+namespace pulp::inspect {
+
+void InspectorOverlay::paint(Canvas& canvas) {
+    if (!active_) return;
+
+    // Phase 6.1 — sample the render-pass manager into the rolling
+    // attribution history once per frame. Cheap, no-op when no RPM is
+    // attached, and de-duplicated against frame_count() so multiple
+    // paints of the same frame don't inflate the history.
+    capture_pass_frame();
+
+    rebuild_flat_tree();
+    // Phase 2 — populate the drift list on the first paint after the
+    // inspector goes active so the drawer is never empty just because
+    // the host forgot to call refresh_drift() explicitly.
+    if (!drift_refreshed_once_) refresh_drift();
+    paint_highlight(canvas);
+    paint_distance_lines(canvas);
+    if (selected_) paint_box_model(canvas, selected_);
+    paint_panel(canvas);
+    // Phase 3c — eyedropper swatch paints above the panel and
+    // highlights so the sampled color is never occluded.
+    paint_eyedropper_cursor(canvas);
+    // Phase 3e — the loupe paints LAST so its magnified grid sits on
+    // top of everything (including the props panel and the eyedropper
+    // swatch), like a physical loupe resting on the design surface.
+    if (zoom_active_) paint_zoom_panel(canvas);
+}
+
+void InspectorOverlay::paint_eyedropper_cursor(Canvas& canvas) {
+    if (!eyedropper_active_) return;
+
+    // Upgrade the sample to an exact framebuffer pixel if the live
+    // Canvas supports readback. Done here (not in the mouse handler)
+    // so the read happens against the fully-composited frame, before
+    // the swatch chrome below would otherwise contaminate the pixel.
+    {
+        Color sampled;
+        if (sample_color_at(eyedropper_cursor_, &canvas, sampled)) {
+            eyedropper_sample_ = sampled;
+            eyedropper_has_sample_ = true;
+        }
+    }
+    if (!eyedropper_has_sample_) return;
+
+    // Swatch + hex readout, offset down-right of the cursor so the
+    // pointer itself never sits on top of the chrome being read.
+    constexpr float kSwatch = 22.0f;   // swatch square side
+    constexpr float kPad    = 4.0f;
+    constexpr float kRowH   = 16.0f;
+    const std::string hex = color_to_hex(eyedropper_sample_);
+
+    canvas.save();
+    canvas.set_font("monospace", kFontSize);
+    float text_w = canvas.measure_text(hex);
+    float box_w = kPad + kSwatch + kPad + text_w + kPad;
+    float box_h = kPad + std::max(kSwatch, kRowH) + kPad;
+    float bx = eyedropper_cursor_.x + 14.0f;
+    float by = eyedropper_cursor_.y + 14.0f;
+
+    // Keep the chrome on-screen near the right / bottom edges.
+    float root_w = root_.bounds().width;
+    float root_h = root_.bounds().height;
+    if (bx + box_w > root_w) bx = eyedropper_cursor_.x - 14.0f - box_w;
+    if (by + box_h > root_h) by = eyedropper_cursor_.y - 14.0f - box_h;
+
+    // Chrome background.
+    canvas.set_fill_color(kEyedropChromeBg);
+    canvas.fill_rounded_rect(bx, by, box_w, box_h, 4.0f);
+
+    // Checkerboard behind the swatch so transparent samples read as
+    // transparent rather than as solid black.
+    float sx = bx + kPad;
+    float sy = by + kPad;
+    const Color kCheckA = Color::rgba(0.75f, 0.75f, 0.75f, 1.0f);
+    const Color kCheckB = Color::rgba(0.55f, 0.55f, 0.55f, 1.0f);
+    constexpr float kCell = kSwatch / 2.0f;
+    for (int cy = 0; cy < 2; ++cy)
+        for (int cx = 0; cx < 2; ++cx) {
+            canvas.set_fill_color(((cx + cy) & 1) ? kCheckB : kCheckA);
+            canvas.fill_rect(sx + cx * kCell, sy + cy * kCell, kCell, kCell);
+        }
+
+    // Sampled color on top of the checkerboard.
+    canvas.set_fill_color(eyedropper_sample_);
+    canvas.fill_rect(sx, sy, kSwatch, kSwatch);
+
+    // Swatch border.
+    canvas.set_stroke_color(kEyedropBorder);
+    canvas.set_line_width(1.0f);
+    canvas.stroke_rect(sx, sy, kSwatch, kSwatch);
+
+    // Hex readout.
+    canvas.set_fill_color(kEyedropText);
+    canvas.fill_text(hex, sx + kSwatch + kPad,
+                     by + box_h / 2.0f + 4.0f);
+
+    canvas.restore();
+}
+
+void InspectorOverlay::paint_highlight(Canvas& canvas) {
+    // Hovered view highlight (blue)
+    if (hovered_ && hovered_ != selected_) {
+        auto r = view_bounds_in_root(hovered_);
+        canvas.set_fill_color(kHighlightFill);
+        canvas.fill_rect(r.x, r.y, r.width, r.height);
+        canvas.set_stroke_color(kHighlightStroke);
+        canvas.set_line_width(1.5f);
+        canvas.stroke_rect(r.x, r.y, r.width, r.height);
+
+        // Tooltip
+        auto type = ViewInspector::type_name(*hovered_);
+        auto label = type + " " + std::to_string(static_cast<int>(r.width))
+                   + "×" + std::to_string(static_cast<int>(r.height));
+        canvas.set_font("monospace", kFontSize);
+        canvas.set_fill_color(kPanelBg);
+        float tw = canvas.measure_text(label);
+        canvas.fill_rounded_rect(r.x, r.y - 18, tw + 8, 16, 3);
+        canvas.set_fill_color(kPanelText);
+        canvas.fill_text(label, r.x + 4, r.y - 5);
+    }
+
+    // Selected view highlight (orange)
+    if (selected_) {
+        auto r = view_bounds_in_root(selected_);
+        canvas.set_fill_color(kSelectedFill);
+        canvas.fill_rect(r.x, r.y, r.width, r.height);
+        canvas.set_stroke_color(kSelectedStroke);
+        canvas.set_line_width(2.0f);
+        canvas.stroke_rect(r.x, r.y, r.width, r.height);
+
+        // Phase 3a — drag handles. Only when dragging mode is on (opt-
+        // in via D key). Four 8×8 filled squares at the corners. The
+        // actively-dragged corner paints in a brighter shade so the
+        // user sees which handle they grabbed even if the cursor
+        // moves slightly off the original target.
+        if (dragging_enabled_) {
+            constexpr float kHandle = 4.0f;  // half-side of 8px box
+            auto paint_handle = [&](float cx, float cy, DragCorner which) {
+                bool active = (active_drag_ == which);
+                canvas.set_fill_color(active
+                    ? Color::rgba(1.0f, 0.7f, 0.2f, 1.0f)   // active = bright orange
+                    : Color::rgba(1.0f, 0.5f, 0.0f, 0.9f)); // idle = same orange as kSelectedStroke
+                canvas.fill_rect(cx - kHandle, cy - kHandle,
+                                 kHandle * 2, kHandle * 2);
+                canvas.set_stroke_color(Color::rgba(0.0f, 0.0f, 0.0f, 0.6f));
+                canvas.set_line_width(1.0f);
+                canvas.stroke_rect(cx - kHandle, cy - kHandle,
+                                   kHandle * 2, kHandle * 2);
+            };
+            paint_handle(r.x,             r.y,              DragCorner::nw);
+            paint_handle(r.x + r.width,   r.y,              DragCorner::ne);
+            paint_handle(r.x,             r.y + r.height,   DragCorner::sw);
+            paint_handle(r.x + r.width,   r.y + r.height,   DragCorner::se);
+        }
+    }
+}
+
+void InspectorOverlay::paint_distance_lines(Canvas& canvas) {
+    // Helper: paint a single distance line + center-to-center px label
+    // between two views. Returns early if either view is missing or the
+    // two are the same.
+    auto paint_one = [&](const View* a_view, const View* b_view) {
+        if (!a_view || !b_view || a_view == b_view) return;
+
+        auto a = view_bounds_in_root(a_view);
+        auto b = view_bounds_in_root(b_view);
+
+        float ax = a.x + a.width / 2;
+        float ay = a.y + a.height / 2;
+        float bx = b.x + b.width / 2;
+        float by = b.y + b.height / 2;
+
+        canvas.set_stroke_color(kDistanceLine);
+        canvas.set_line_width(1.0f);
+        canvas.stroke_line(ax, ay, bx, by);
+
+        // Distance label
+        float dx = bx - ax;
+        float dy = by - ay;
+        float dist = std::sqrt(dx * dx + dy * dy);
+        auto label = std::to_string(static_cast<int>(dist)) + "px";
+        float mx = (ax + bx) / 2;
+        float my = (ay + by) / 2;
+
+        canvas.set_font("monospace", kFontSize);
+        canvas.set_fill_color(kDistanceLine);
+        float tw = canvas.measure_text(label);
+        canvas.fill_rounded_rect(mx - tw / 2 - 4, my - 8, tw + 8, 16, 3);
+        canvas.set_fill_color(Color::rgba(1, 1, 1, 1));
+        canvas.fill_text(label, mx - tw / 2, my + 4);
+    };
+
+    // Existing: Alt+click sticky distance-anchor mode
+    paint_one(distance_anchor_, selected_);
+
+    // Phase 3f: Alt-hover sibling distance (Figma-style spacing reveal).
+    // While Alt is held during hover, dynamically paint a line from the
+    // current selection to the view under the cursor. The two modes can
+    // coexist — sticky anchor + live hover — for richer measurement.
+    if (alt_hover_target_ && selected_ &&
+        alt_hover_target_ != distance_anchor_) {
+        paint_one(selected_, alt_hover_target_);
+    }
+}
+
+void InspectorOverlay::paint_box_model(Canvas& canvas, const View* v) {
+    if (!v || !v->parent()) return;
+
+    auto r = view_bounds_in_root(v);
+    auto& f = v->flex();
+
+    // Padding (green, inside the view)
+    float pt = f.padding_top >= 0 ? f.padding_top : f.padding;
+    float pr = f.padding_right >= 0 ? f.padding_right : f.padding;
+    float pb = f.padding_bottom >= 0 ? f.padding_bottom : f.padding;
+    float pl = f.padding_left >= 0 ? f.padding_left : f.padding;
+    if (pt > 0 || pr > 0 || pb > 0 || pl > 0) {
+        canvas.set_fill_color(kPaddingColor);
+        if (pt > 0) canvas.fill_rect(r.x, r.y, r.width, pt);
+        if (pb > 0) canvas.fill_rect(r.x, r.y + r.height - pb, r.width, pb);
+        if (pl > 0) canvas.fill_rect(r.x, r.y + pt, pl, r.height - pt - pb);
+        if (pr > 0) canvas.fill_rect(r.x + r.width - pr, r.y + pt, pr, r.height - pt - pb);
+    }
+
+    // Margin (orange, outside the view)
+    float mt = f.margin_t();
+    float mr_ = f.margin_r();
+    float mb = f.margin_b();
+    float ml = f.margin_l();
+    if (mt > 0 || mr_ > 0 || mb > 0 || ml > 0) {
+        canvas.set_fill_color(kMarginColor);
+        if (mt > 0) canvas.fill_rect(r.x - ml, r.y - mt, r.width + ml + mr_, mt);
+        if (mb > 0) canvas.fill_rect(r.x - ml, r.y + r.height, r.width + ml + mr_, mb);
+        if (ml > 0) canvas.fill_rect(r.x - ml, r.y, ml, r.height);
+        if (mr_ > 0) canvas.fill_rect(r.x + r.width, r.y, mr_, r.height);
+    }
+
+    // Distance to parent
+    auto parent_r = view_bounds_in_root(v->parent());
+    float dist_top = r.y - parent_r.y;
+    float dist_left = r.x - parent_r.x;
+    float dist_bottom = (parent_r.y + parent_r.height) - (r.y + r.height);
+    float dist_right = (parent_r.x + parent_r.width) - (r.x + r.width);
+
+    canvas.set_font("monospace", 9.0f);
+    canvas.set_fill_color(kDistanceLine);
+    if (dist_top > 2) canvas.fill_text(std::to_string(static_cast<int>(dist_top)), r.x + r.width / 2, r.y - 3);
+    if (dist_left > 2) canvas.fill_text(std::to_string(static_cast<int>(dist_left)), r.x - 20, r.y + r.height / 2);
+    if (dist_bottom > 2) canvas.fill_text(std::to_string(static_cast<int>(dist_bottom)), r.x + r.width / 2, r.y + r.height + 10);
+    if (dist_right > 2) canvas.fill_text(std::to_string(static_cast<int>(dist_right)), r.x + r.width + 3, r.y + r.height / 2);
+}
+
+void InspectorOverlay::paint_panel(Canvas& canvas) {
+    float root_w = root_.bounds().width;
+    float root_h = root_.bounds().height;
+    float panel_x = root_w - panel_width_;
+
+    // Panel background
+    canvas.save();
+    canvas.set_fill_color(kPanelBg);
+    canvas.fill_rect(panel_x, 0, panel_width_, root_h);
+
+    // Divider line
+    canvas.set_stroke_color(Color::rgba(0.3f, 0.3f, 0.35f, 1.0f));
+    canvas.set_line_width(1.0f);
+    canvas.stroke_line(panel_x, 0, panel_x, root_h);
+
+    float stats_y = root_h - kStatsBarHeight;
+
+    // Helper: paint the middle "props" region. Phase 6.1 — when the
+    // per-pass attribution viewer is toggled on (P-key), it takes over
+    // this region instead of the property panel; the tree section above
+    // is untouched so the user keeps navigation context. Phase 5.2 —
+    // the reconciliation tab (R-key) takes over the same region with
+    // the same discipline. Phase 6.2 — the texture-atlas viewer (A-key)
+    // is the third tab to claim this region. Precedence when multiple
+    // are toggled, oldest surface wins: pass viewer > reconciliation >
+    // atlas viewer > props. The losers' cached row counts are reset so
+    // reconcile_row_count() / atlas_row_count() never report a stale
+    // layout for a tab that did not paint this frame.
+    auto paint_middle = [&](float x, float y, float w, float h) {
+        if (pass_viewer_enabled_) {
+            reconcile_rows_.clear();
+            atlas_row_count_ = 0;
+            paint_pass_attribution(canvas, x, y, w, h);
+        } else if (reconcile_tab_visible_) {
+            atlas_row_count_ = 0;
+            paint_reconcile_tab(canvas, x, y, w, h);
+        } else if (atlas_viewer_visible_) {
+            reconcile_rows_.clear();
+            paint_atlas_tab(canvas, x, y, w, h);
+        } else {
+            reconcile_rows_.clear();
+            atlas_row_count_ = 0;
+            paint_props_section(canvas, x, y, w, h);
+        }
+    };
+
+    if (tweaks_panel_visible_) {
+        // Phase 2.5 layout: tree (top third), props (middle third),
+        // tweaks management panel (bottom third). When the panel is
+        // hidden the legacy two-section layout is used (below).
+        float section_h = (stats_y) / 3.0f;
+
+        float cursor_y = 4.0f;
+        paint_tree_section(canvas, panel_x + 8, 4, panel_width_ - 16, cursor_y);
+
+        float props_y = section_h;
+        canvas.set_stroke_color(Color::rgba(0.3f, 0.3f, 0.35f, 0.5f));
+        canvas.stroke_line(panel_x + 8, props_y, root_w - 8, props_y);
+
+        // Phase 2 — drift drawer sits directly under the tree divider.
+        // Paints nothing (and returns 0) when there is no drift, so the
+        // props section is unaffected on the happy path.
+        float drift_h = paint_drift_drawer(canvas, panel_x + 8, props_y + 4,
+                                           panel_width_ - 16);
+        paint_middle(panel_x + 8, props_y + 4 + drift_h,
+                     panel_width_ - 16, section_h - 8 - drift_h);
+
+        float tweaks_y = section_h * 2.0f;
+        canvas.set_stroke_color(Color::rgba(0.3f, 0.3f, 0.35f, 0.5f));
+        canvas.stroke_line(panel_x + 8, tweaks_y, root_w - 8, tweaks_y);
+        paint_tweaks_section(canvas, panel_x + 8, tweaks_y + 4,
+                             panel_width_ - 16, stats_y - tweaks_y - 8);
+    } else {
+        // Legacy two-section layout (pre-2.5).
+        tweak_rows_.clear();
+        float tree_height = root_h * 0.5f;
+        float cursor_y = 4.0f;
+        paint_tree_section(canvas, panel_x + 8, 4, panel_width_ - 16, cursor_y);
+
+        float props_y = tree_height;
+        canvas.set_stroke_color(Color::rgba(0.3f, 0.3f, 0.35f, 0.5f));
+        canvas.stroke_line(panel_x + 8, props_y, root_w - 8, props_y);
+
+        // Phase 2 — drift drawer sits directly under the tree divider.
+        // Paints nothing (and returns 0) when there is no drift, so the
+        // props section is unaffected on the happy path.
+        float drift_h = paint_drift_drawer(canvas, panel_x + 8, props_y + 4,
+                                           panel_width_ - 16);
+        paint_middle(panel_x + 8, props_y + 4 + drift_h,
+                     panel_width_ - 16, stats_y - props_y - 8 - drift_h);
+    }
+
+    // Stats bar (bottom)
+    paint_stats_bar(canvas, panel_x, stats_y, panel_width_);
+
+    canvas.restore();
+}
+
+void InspectorOverlay::paint_tree_section(Canvas& canvas, float x, float y, float w, float& cursor_y) {
+    canvas.set_font("monospace", kFontSize);
+    float tree_height = root_.bounds().height * 0.5f;
+
+    for (auto& item : flat_tree_) {
+        float row_y = y + cursor_y - tree_scroll_y_;
+        if (row_y < y - kRowHeight || row_y > y + tree_height) {
+            cursor_y += kRowHeight;
+            continue;
+        }
+
+        float indent = item.depth * kIndent;
+
+        // Highlight selected row
+        if (item.view == selected_) {
+            canvas.set_fill_color(kTreeSelected);
+            canvas.fill_rect(x, row_y, w, kRowHeight);
+        } else if (item.view == hovered_) {
+            canvas.set_fill_color(kPanelHighlight);
+            canvas.fill_rect(x, row_y, w, kRowHeight);
+        }
+
+        // Collapse indicator
+        if (item.view->child_count() > 0) {
+            canvas.set_fill_color(kPanelDim);
+            auto indicator = collapsed_.count(item.view) ? "\xe2\x96\xb6" : "\xe2\x96\xbc";
+            canvas.fill_text(indicator, x + indent, row_y + 14);
+        }
+
+        // Type name + optional ID
+        auto type = ViewInspector::type_name(*item.view);
+        auto id = item.view->id();
+        std::string label = type;
+        if (!id.empty()) label += " #" + id;
+
+        canvas.set_fill_color(kPanelText);
+        canvas.fill_text(label, x + indent + 14, row_y + 14);
+
+        cursor_y += kRowHeight;
+    }
+}
+
+void InspectorOverlay::paint_props_section(Canvas& canvas, float x, float y, float w, float h) {
+    // Clear last frame's editable-field rects — repainted as we go.
+    // The mouse handler reads this list during the SAME frame the user
+    // clicked (paint runs on the UI thread before input dispatch).
+    editable_fields_.clear();
+
+    if (!selected_) {
+        canvas.set_fill_color(kPanelDim);
+        canvas.set_font("monospace", kFontSize);
+        canvas.fill_text("Click a view to inspect", x, y + 16);
+        return;
+    }
+
+    canvas.set_font("monospace", kFontSize);
+    float line_y = y + 4;
+    float line_h = 15.0f;
+
+    // Phase 0b PR-C-2 — dot indicator for properties with local tweaks.
+    // When the TweakStore (PR-A) has an entry for this view's anchor at
+    // the given dotted path, paint a small orange dot in the gutter to
+    // the LEFT of the label. The label/value text remains untouched so
+    // the row still reads cleanly without color. The dot stays small
+    // (3px radius) so the panel doesn't pulse with visual noise when
+    // many tweaks are active. Designed to live next to (not over) the
+    // existing label column at x.
+    auto has_tweak = [&](std::string_view path) -> bool {
+        if (!tweak_store_) return false;
+        const auto& anchor = selected_->anchor_id();
+        if (anchor.empty()) return false;
+        if (tweak_store_->is_bypassed(anchor, path)) return false;
+        return tweak_store_->lookup(anchor, path).has_value();
+    };
+
+    auto draw_label = [&](const std::string& label, const std::string& value,
+                          std::string_view tweak_path = {}) {
+        if (line_y > y + h) return;
+        if (!tweak_path.empty() && has_tweak(tweak_path)) {
+            // Orange dot indicator — same hue as the kSelectedStroke
+            // selection color so users associate it visually with
+            // "this is a Pulp-owned modification."
+            canvas.set_fill_color(Color::rgba(1.0f, 0.5f, 0.0f, 0.9f));
+            canvas.fill_circle(x - 6, line_y + 8, 3);
+        }
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text(label, x, line_y + 11);
+        canvas.set_fill_color(kPanelText);
+        canvas.fill_text(value, x + 80, line_y + 11);
+        line_y += line_h;
+    };
+
+    auto draw_heading = [&](const std::string& text) {
+        if (line_y > y + h) return;
+        line_y += 4;
+        canvas.set_fill_color(kHighlightStroke);
+        canvas.fill_text(text, x, line_y + 11);
+        line_y += line_h;
+    };
+
+    // Phase 3b — emit one editable numeric row. Reserves a click-hit
+    // rect in editable_fields_ keyed by dotted property path. If this
+    // is the row currently being edited, draws the live edit_buffer_
+    // text with a thin underline + caret instead of the static value.
+    auto draw_editable = [&](const std::string& label,
+                             const std::string& field_path,
+                             float value,
+                             const std::string& formatted_value) {
+        if (line_y > y + h) return;
+        // Field hit-rect — covers the entire value area, not the label,
+        // so a click on "padding" text doesn't accidentally edit. Width
+        // extends to the right edge of the panel section so the click
+        // target is generous (Fitts's law).
+        float value_x = x + 80;
+        float value_w = (x + w) - value_x;
+        Rect hit{value_x, line_y, value_w, line_h};
+        editable_fields_.push_back({field_path, hit, value});
+
+        // Phase 0b PR-C-2 — dot indicator coexists with Phase 3b edit
+        // mode. The tweak-path here matches the field_path the editable
+        // emits on commit, so an edit immediately shows a dot next time
+        // the panel paints (and clears when bypassed). Drawn before the
+        // label so the label/value text remains untouched.
+        if (!field_path.empty() && has_tweak(field_path)) {
+            canvas.set_fill_color(Color::rgba(1.0f, 0.5f, 0.0f, 0.9f));
+            canvas.fill_circle(x - 6, line_y + 8, 3);
+        }
+
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text(label, x, line_y + 11);
+
+        const bool editing_this = (editing_field_ == field_path);
+        if (editing_this) {
+            // Edit-mode background tint + underline + caret.
+            canvas.set_fill_color(kFieldEditBg);
+            canvas.fill_rect(value_x - 2, line_y, value_w + 4, line_h);
+
+            // Show the live buffer.
+            canvas.set_fill_color(kPanelText);
+            const std::string& buf = edit_buffer_;
+            canvas.fill_text(buf, value_x, line_y + 11);
+
+            // Caret — width of "0" is a fine monospace approximation
+            // since we set font("monospace", kFontSize). Measure the
+            // prefix up to caret_pos for the X offset.
+            std::string prefix = buf.substr(0, std::min(edit_caret_pos_, buf.size()));
+            float caret_x = value_x + canvas.measure_text(prefix);
+            canvas.set_stroke_color(kFieldEditCaret);
+            canvas.set_line_width(1.0f);
+            canvas.stroke_line(caret_x, line_y + 2, caret_x, line_y + line_h - 2);
+
+            // Underline along the whole value area.
+            canvas.set_stroke_color(kFieldEditUnder);
+            canvas.set_line_width(1.0f);
+            canvas.stroke_line(value_x, line_y + line_h - 1,
+                               value_x + value_w, line_y + line_h - 1);
+        } else {
+            // Non-edit state: just the value text. The hit-rect is
+            // invisible — Phase 3b intentionally keeps the chrome
+            // minimal; a hover-cursor hint is future-work for the
+            // platform host (see editable_field_at()).
+            canvas.set_fill_color(kPanelText);
+            canvas.fill_text(formatted_value, value_x, line_y + 11);
+        }
+
+        line_y += line_h;
+    };
+
+    // Type and ID
+    auto type = ViewInspector::type_name(*selected_);
+    draw_heading(type + (selected_->id().empty() ? "" : " #" + selected_->id()));
+
+    // Phase 5.1 — authored-source row. When the selected view carries a
+    // `__source` provenance record (set via the JS bridge's setSource()
+    // for JSX-imported views), show "file:line" and a hint that the J
+    // hotkey jumps to it. Hidden entirely for non-imported views so the
+    // panel stays uncluttered.
+    if (selected_->has_source_loc()) {
+        const auto& loc = selected_->source_loc();
+        std::string where = loc.file;
+        if (loc.line > 0) where += ":" + std::to_string(loc.line);
+        draw_label("source", where);
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text("(press J to open in editor)", x, line_y + 11);
+        line_y += line_h;
+    }
+
+    // Bounds (informational — bounds is layout OUTPUT, not editable.
+    // Edits go through flex inputs below.)
+    auto r = selected_->bounds();
+    auto abs = view_bounds_in_root(selected_);
+    draw_label("bounds", std::to_string(static_cast<int>(r.x)) + ", " +
+               std::to_string(static_cast<int>(r.y)) + ", " +
+               std::to_string(static_cast<int>(r.width)) + " × " +
+               std::to_string(static_cast<int>(r.height)));
+    draw_label("absolute", std::to_string(static_cast<int>(abs.x)) + ", " +
+               std::to_string(static_cast<int>(abs.y)));
+
+    // Visibility (not editable in Phase 3b — Phase 0b PR-C-2 adds dot)
+    draw_label("visible", selected_->visible() ? "true" : "false", "paint.visible");
+
+    // Phase 3b editable: opacity (always present, default 1.0).
+    // Phase 0b PR-C-2: draw_editable now also paints the tweak dot when
+    // style.opacity has a TweakStore entry.
+    {
+        float op = selected_->opacity();
+        std::ostringstream oss;
+        oss << std::fixed << std::setprecision(2) << op;
+        draw_editable("opacity", "style.opacity", op, oss.str());
+    }
+
+    // Flex
+    auto& f = selected_->flex();
+    draw_heading("Layout");
+
+    auto dir_str = [](FlexDirection d) -> std::string {
+        switch (d) {
+            case FlexDirection::row: return "row";
+            case FlexDirection::column: return "column";
+        }
+        return "?";
+    };
+    draw_label("direction", dir_str(f.direction), "layout.direction");
+
+    if (f.flex_grow > 0) draw_label("grow", std::to_string(f.flex_grow), "layout.grow");
+    if (f.flex_shrink != 1.0f) draw_label("shrink", std::to_string(f.flex_shrink), "layout.shrink");
+    if (f.gap > 0) draw_label("gap", std::to_string(static_cast<int>(f.gap)), "layout.gap");
+
+    // Phase 3b editable: width / height — uses preferred_width /
+    // preferred_height which are the Yoga flex INPUTS (Codex Phase 3
+    // correction: set_bounds is an output that Yoga overwrites).
+    {
+        std::ostringstream oss;
+        oss << static_cast<int>(f.preferred_width);
+        draw_editable("width", "layout.width", f.preferred_width, oss.str());
+    }
+    {
+        std::ostringstream oss;
+        oss << static_cast<int>(f.preferred_height);
+        draw_editable("height", "layout.height", f.preferred_height, oss.str());
+    }
+
+    // Phase 3b editable: padding (uniform). Per-side editing is a
+    // future enhancement — exposed via the per-side rows below when
+    // per-side overrides are already in use, otherwise the uniform
+    // single field is the clean common case.
+    {
+        std::ostringstream oss;
+        oss << static_cast<int>(f.padding);
+        draw_editable("padding", "layout.padding", f.padding, oss.str());
+    }
+
+    // Phase 3b editable: margin (uniform).
+    {
+        std::ostringstream oss;
+        oss << static_cast<int>(f.margin);
+        draw_editable("margin", "layout.margin", f.margin, oss.str());
+    }
+
+    float pt2 = f.padding_top >= 0 ? f.padding_top : f.padding;
+    float pr2 = f.padding_right >= 0 ? f.padding_right : f.padding;
+    float pb2 = f.padding_bottom >= 0 ? f.padding_bottom : f.padding;
+    float pl2 = f.padding_left >= 0 ? f.padding_left : f.padding;
+    if ((pt2 != f.padding) || (pr2 != f.padding) ||
+        (pb2 != f.padding) || (pl2 != f.padding)) {
+        draw_label("padding (sides)",
+                   std::to_string(static_cast<int>(pt2)) + " " +
+                   std::to_string(static_cast<int>(pr2)) + " " +
+                   std::to_string(static_cast<int>(pb2)) + " " +
+                   std::to_string(static_cast<int>(pl2)),
+                   "layout.padding");
+    }
+
+    // Theme colors (first 5)
+    auto& theme = selected_->theme();
+    if (!theme.colors.empty()) {
+        draw_heading("Theme Colors");
+        int shown = 0;
+        for (auto& [name, color] : theme.colors) {
+            if (shown >= 5) {
+                draw_label("", "... +" + std::to_string(theme.colors.size() - 5) + " more");
+                break;
+            }
+            // Color swatch
+            if (line_y <= y + h) {
+                canvas.set_fill_color(color);
+                canvas.fill_rounded_rect(x, line_y + 3, 10, 10, 2);
+                canvas.set_fill_color(kPanelDim);
+                canvas.fill_text(name, x + 16, line_y + 11);
+                line_y += line_h;
+            }
+            ++shown;
+        }
+    }
+}
+
+void InspectorOverlay::paint_stats_bar(Canvas& canvas, float x, float y, float w) {
+    canvas.set_fill_color(kStatsBg);
+    canvas.fill_rect(x, y, w, kStatsBarHeight);
+
+    canvas.set_font("monospace", 10.0f);
+
+    if (rpm_) {
+        float frame_ms = rpm_->total_time_ms();
+        float budget = rpm_->budget();
+        bool over = rpm_->over_budget();
+
+        std::ostringstream ss;
+        ss << std::fixed << std::setprecision(1) << frame_ms << "ms";
+        if (budget > 0) {
+            int fps = static_cast<int>(1000.0f / std::max(frame_ms, 0.1f));
+            ss << "  " << fps << "fps";
+        }
+
+        canvas.set_fill_color(over ? kStatsWarn : kStatsText);
+        canvas.fill_text(ss.str(), x + 8, y + 16);
+
+        // Pass breakdown
+        auto& passes = rpm_->passes();
+        if (!passes.empty()) {
+            std::string pass_info;
+            for (auto& p : passes) {
+                if (!pass_info.empty()) pass_info += " | ";
+                pass_info += std::to_string(p.draw_calls) + "dc";
+            }
+            canvas.set_fill_color(kPanelDim);
+            canvas.fill_text(pass_info, x + 140, y + 16);
+        }
+    } else {
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text("No render stats", x + 8, y + 16);
+    }
+
+    // Phase 3c — active-mode hint. Drag (D) and eyedropper (E) are
+    // mutually informative; show whichever is armed so the user
+    // remembers a non-default canvas mode is in effect. Placed at the
+    // bar midpoint so it never overlaps the left-aligned frame-time
+    // readout or the right-aligned view count.
+    if (eyedropper_active_ || dragging_enabled_) {
+        canvas.set_fill_color(kFieldEditCaret);
+        const char* mode = eyedropper_active_ ? "\xe2\x97\x89 eyedropper"
+                                              : "\xe2\x97\x89 drag";
+        canvas.fill_text(mode, x + w * 0.5f - 24.0f, y + 16);
+    }
+
+    // View count
+    auto count = ViewInspector::count_views(root_);
+    canvas.set_fill_color(kPanelDim);
+    canvas.fill_text(std::to_string(count) + " views", x + w - 60, y + 16);
+}
+
+// Phase 6.1 — the per-pass GPU/render attribution viewer
+// (capture_pass_frame / pass_attribution / paint_pass_attribution) is
+// defined in inspector_overlay_pass_viewer.cpp.
+
+// ── Phase 2.5 — Tweak management panel (Photoshop-layers style) ─────────────
+//
+// Lists every tweak in the attached TweakStore, grouped by anchor.
+// Each tweak is a "layer" with three per-tweak controls:
+//   eye   — bypass toggle (filled when active, hollow when bypassed)
+//   lock  — protect from bulk-clear / reimport
+//   trash — delete the tweak
+// The row hit-rects are stashed in tweak_rows_ so the same-frame mouse
+// handler can resolve a click. Anchor headers carry an abbreviated id;
+// each row shows the dotted property path + a compact value preview.
+
+namespace {
+
+// Compact value preview for a tweak — keeps the row narrow. Numbers
+// print without trailing-zero noise; strings clip to 16 chars.
+std::string preview_value(const choc::value::Value& v) {
+    if (v.isString()) {
+        auto s = std::string(v.getString());
+        if (s.size() > 16) s = s.substr(0, 15) + "\xe2\x80\xa6";
+        return "\"" + s + "\"";
+    }
+    if (v.isBool()) return v.getBool() ? "true" : "false";
+    if (v.isInt32() || v.isInt64())
+        return std::to_string(v.getWithDefault<int64_t>(0));
+    if (v.isFloat32() || v.isFloat64()) {
+        std::ostringstream oss;
+        oss << v.getWithDefault<double>(0.0);
+        return oss.str();
+    }
+    if (v.isObject()) return "{\xe2\x80\xa6}";
+    if (v.isArray()) return "[\xe2\x80\xa6]";
+    return "?";
+}
+
+// Abbreviate an anchor id for the header — keep the tail (most
+// distinctive) but cap total width so the header never overflows.
+std::string abbreviate_anchor(const std::string& id) {
+    constexpr std::size_t kMax = 22;
+    if (id.size() <= kMax) return id;
+    return "\xe2\x80\xa6" + id.substr(id.size() - (kMax - 1));
+}
+
+}  // namespace
+
+void InspectorOverlay::paint_tweaks_section(Canvas& canvas, float x, float y,
+                                            float w, float h) {
+    // Repopulated every frame — the mouse handler reads it on the same
+    // frame the user clicked (paint runs before input dispatch).
+    tweak_rows_.clear();
+
+    canvas.set_font("monospace", kFontSize);
+
+    // Section heading.
+    canvas.set_fill_color(kHighlightStroke);
+    canvas.fill_text("Tweaks", x, y + 11);
+
+    if (!tweak_store_) {
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text("No tweak store attached", x, y + 11 + kRowHeight);
+        return;
+    }
+
+    auto records = tweak_store_->list_tweaks();
+    {
+        std::ostringstream count_oss;
+        count_oss << records.size() << (records.size() == 1 ? " tweak" : " tweaks");
+        canvas.set_fill_color(kPanelDim);
+        float cw = canvas.measure_text(count_oss.str());
+        canvas.fill_text(count_oss.str(), x + w - cw, y + 11);
+    }
+
+    if (records.empty()) {
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text("No tweaks recorded", x, y + 11 + kRowHeight);
+        return;
+    }
+
+    // Group records by anchor — stable insertion order within an
+    // anchor (list_tweaks() preserves it), anchors sorted so the panel
+    // doesn't reshuffle across frames.
+    std::vector<std::string> anchor_order;
+    std::unordered_map<std::string, std::vector<const TweakStore::Record*>> grouped;
+    for (auto& rec : records) {
+        auto it = grouped.find(rec.anchor_id);
+        if (it == grouped.end()) {
+            grouped.emplace(rec.anchor_id,
+                            std::vector<const TweakStore::Record*>{&rec});
+            anchor_order.push_back(rec.anchor_id);
+        } else {
+            it->second.push_back(&rec);
+        }
+    }
+    std::sort(anchor_order.begin(), anchor_order.end());
+
+    // Clip the scroll region so rows don't bleed into the stats bar.
+    canvas.save();
+    float list_top = y + kRowHeight;
+    canvas.clip_rect(x, list_top, w, h - kRowHeight);
+
+    float row_y = list_top - tweaks_scroll_y_;
+    constexpr float kIconSize = 14.0f;
+    constexpr float kIconGap = 4.0f;
+    // Three icons stacked at the right edge of the row.
+    float icons_w = 3.0f * kIconSize + 2.0f * kIconGap;
+
+    for (auto& anchor : anchor_order) {
+        bool anchor_locked = tweak_store_->is_locked(anchor);
+
+        // Anchor header row.
+        if (row_y > list_top - kRowHeight && row_y < y + h) {
+            canvas.set_fill_color(kPanelHighlight);
+            canvas.fill_rect(x, row_y, w, kRowHeight);
+            canvas.set_fill_color(kHighlightStroke);
+            std::string hdr = abbreviate_anchor(anchor);
+            if (anchor_locked) hdr = "\xf0\x9f\x94\x92 " + hdr;  // 🔒
+            canvas.fill_text(hdr, x + 2, row_y + 14);
+        }
+        row_y += kRowHeight;
+
+        for (const auto* rec : grouped[anchor]) {
+            bool visible = row_y > list_top - kRowHeight && row_y < y + h;
+            bool bypassed = tweak_store_->is_bypassed(rec->anchor_id,
+                                                      rec->property_path);
+
+            // Always record the hit-rects, even off-screen rows, so
+            // tweak_row_count() is the true total; off-screen rects
+            // simply never get clicked.
+            float icons_x = x + w - icons_w;
+            TweakRow hr;
+            hr.anchor_id = rec->anchor_id;
+            hr.property_path = rec->property_path;
+            hr.bypass_icon = {icons_x, row_y + 3, kIconSize, kIconSize};
+            hr.lock_icon   = {icons_x + kIconSize + kIconGap, row_y + 3,
+                              kIconSize, kIconSize};
+            hr.delete_icon = {icons_x + 2 * (kIconSize + kIconGap), row_y + 3,
+                              kIconSize, kIconSize};
+            tweak_rows_.push_back(hr);
+
+            if (visible) {
+                // Bypassed rows render dimmed to read like a hidden
+                // Photoshop layer.
+                const Color& path_col = bypassed ? kPanelDim : kPanelText;
+
+                // Property path (indented under the anchor header).
+                std::string label = rec->property_path;
+                float max_label_w = (icons_x - kIconGap) - (x + kIndent);
+                while (label.size() > 4 &&
+                       canvas.measure_text(label + " = ") > max_label_w)
+                    label = label.substr(0, label.size() - 1);
+                canvas.set_fill_color(path_col);
+                canvas.fill_text(label, x + kIndent, row_y + 14);
+
+                // Value preview.
+                std::string val = preview_value(rec->value);
+                canvas.set_fill_color(kPanelDim);
+                float label_w = canvas.measure_text(label + " ");
+                canvas.fill_text("= " + val, x + kIndent + label_w, row_y + 14);
+
+                // ── Icon: bypass (eye) ──────────────────────────────
+                // Filled circle = visible/applied; hollow = bypassed.
+                {
+                    auto& r = hr.bypass_icon;
+                    float cx = r.x + r.width / 2, cy = r.y + r.height / 2;
+                    if (bypassed) {
+                        canvas.set_stroke_color(kPanelDim);
+                        canvas.set_line_width(1.5f);
+                        canvas.stroke_line(r.x, r.y + r.height,
+                                           r.x + r.width, r.y);
+                        canvas.set_stroke_color(kPanelDim);
+                        canvas.stroke_rect(r.x, r.y, r.width, r.height);
+                    } else {
+                        canvas.set_fill_color(kHighlightStroke);
+                        canvas.fill_circle(cx, cy, r.width / 2 - 1);
+                    }
+                }
+                // ── Icon: lock ──────────────────────────────────────
+                {
+                    auto& r = hr.lock_icon;
+                    Color lock_col = anchor_locked
+                        ? Color::rgba(1.0f, 0.78f, 0.2f, 1.0f)
+                        : kPanelDim;
+                    // Shackle (arc approximated by a rounded rect) +
+                    // body box.
+                    canvas.set_stroke_color(lock_col);
+                    canvas.set_line_width(1.5f);
+                    canvas.stroke_rect(r.x + 3, r.y + 1, r.width - 6,
+                                       r.height / 2);
+                    canvas.set_fill_color(lock_col);
+                    canvas.fill_rounded_rect(r.x + 1, r.y + r.height / 2 - 1,
+                                             r.width - 2, r.height / 2, 2);
+                }
+                // ── Icon: delete (trash) ────────────────────────────
+                {
+                    auto& r = hr.delete_icon;
+                    canvas.set_stroke_color(
+                        Color::rgba(1.0f, 0.4f, 0.35f, 1.0f));
+                    canvas.set_line_width(1.5f);
+                    // Lid.
+                    canvas.stroke_line(r.x + 1, r.y + 3,
+                                       r.x + r.width - 1, r.y + 3);
+                    // Body box.
+                    canvas.stroke_rect(r.x + 2, r.y + 3, r.width - 4,
+                                       r.height - 4);
+                    // Handle.
+                    canvas.stroke_line(r.x + r.width / 2 - 2, r.y + 1,
+                                       r.x + r.width / 2 + 2, r.y + 1);
+                }
+            }
+            row_y += kRowHeight;
+        }
+    }
+
+    canvas.restore();
+}
+
+InspectorOverlay::TweakAction
+InspectorOverlay::tweak_action_at(Point p, std::size_t& out_row) const {
+    auto hit = [&](const Rect& r) {
+        return p.x >= r.x && p.x <= r.x + r.width &&
+               p.y >= r.y && p.y <= r.y + r.height;
+    };
+    for (std::size_t i = 0; i < tweak_rows_.size(); ++i) {
+        const auto& row = tweak_rows_[i];
+        if (hit(row.bypass_icon)) { out_row = i; return TweakAction::bypass; }
+        if (hit(row.lock_icon))   { out_row = i; return TweakAction::lock; }
+        if (hit(row.delete_icon)) { out_row = i; return TweakAction::remove; }
+    }
+    return TweakAction::none;
+}
+
+// ── Phase 5.2 — Reconciliation tab ──────────────────────────────────────────
+//
+// A read-only report tab (R-key) showing, per tweak, whether the edit
+// will survive a fresh design re-import. It classifies every stored
+// tweak via reconcile_report() and renders one row each with a
+// color-coded status badge. Like the Phase 6.1 pass viewer it takes
+// over the property-panel region; unlike the tweak management panel it
+// has no interactive controls — it is purely informational, so there
+// are no hit-rects to record.
+
+void InspectorOverlay::paint_reconcile_tab(Canvas& canvas, float x, float y,
+                                           float w, float h) {
+    // Status badge colors — green = reconciled (safe), amber = drift
+    // (runtime-only), red = unresolvable (orphaned). The amber/red pair
+    // matches the Phase 2 drift drawer so the two surfaces read
+    // consistently.
+    const Color kLockedColor = Color::rgba(0.35f, 0.85f, 0.45f, 1.0f);
+    const Color kDriftColor  = Color::rgba(0.95f, 0.65f, 0.25f, 1.0f);
+    const Color kUnresColor  = Color::rgba(0.95f, 0.40f, 0.38f, 1.0f);
+
+    canvas.set_font("monospace", kFontSize);
+
+    // Section heading.
+    canvas.set_fill_color(kHighlightStroke);
+    canvas.fill_text("Reconcile", x, y + 11);
+
+    // Recompute the report fresh every frame — it is a cheap O(tweaks)
+    // pass and the live tree / lock state may have changed since the
+    // last paint. reconcile_rows_ caches the result for the row count.
+    auto report = reconcile_report();
+    reconcile_rows_ = report.rows;
+
+    if (!tweak_store_) {
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text("No tweak store attached", x, y + 11 + kRowHeight);
+        return;
+    }
+
+    // Summary line: per-status counts, right-aligned in the header.
+    {
+        std::ostringstream summary;
+        summary << report.locked_count << " locked  "
+                << report.drifted_count << " drift  "
+                << report.unresolvable_count << " unresolved";
+        canvas.set_fill_color(kPanelDim);
+        float sw = canvas.measure_text(summary.str());
+        canvas.fill_text(summary.str(), x + w - sw, y + 11);
+    }
+
+    if (report.rows.empty()) {
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text("No tweaks to reconcile", x, y + 11 + kRowHeight);
+        return;
+    }
+
+    // Clip the scroll region so rows don't bleed past the section.
+    canvas.save();
+    float list_top = y + kRowHeight;
+    canvas.clip_rect(x, list_top, w, h - kRowHeight);
+
+    constexpr float kBadgeW = 16.0f;  // status pip diameter
+    float row_y = list_top - reconcile_scroll_y_;
+
+    for (const auto& row : report.rows) {
+        const bool visible = row_y > list_top - kRowHeight && row_y < y + h;
+        if (visible) {
+            Color badge;
+            const char* tag;
+            switch (row.status) {
+                case ReconcileStatus::locked_to_source:
+                    badge = kLockedColor; tag = "lock"; break;
+                case ReconcileStatus::drifted:
+                    badge = kDriftColor;  tag = "drift"; break;
+                case ReconcileStatus::unresolvable:
+                default:
+                    badge = kUnresColor;  tag = "orphan"; break;
+            }
+
+            // Status pip at the left edge.
+            canvas.set_fill_color(badge);
+            canvas.fill_circle(x + 5.0f, row_y + kRowHeight / 2.0f, 4.0f);
+
+            // Anchor + property path, dimmed for unresolvable rows so
+            // an orphaned tweak reads as "inactive" at a glance.
+            const bool orphan = row.status == ReconcileStatus::unresolvable;
+            std::string label = abbreviate_anchor(row.anchor_id) + "  " +
+                                 row.property_path;
+            float text_x = x + kBadgeW;
+            float tag_w = canvas.measure_text(tag) + 6.0f;
+            float max_label_w = (x + w - tag_w) - text_x;
+            while (label.size() > 4 &&
+                   canvas.measure_text(label) > max_label_w)
+                label = label.substr(0, label.size() - 1);
+            canvas.set_fill_color(orphan ? kPanelDim : kPanelText);
+            canvas.fill_text(label, text_x, row_y + 14);
+
+            // Status tag, right-aligned, in the badge color.
+            canvas.set_fill_color(badge);
+            canvas.fill_text(tag, x + w - canvas.measure_text(tag),
+                             row_y + 14);
+        }
+        row_y += kRowHeight;
+    }
+
+    canvas.restore();
+}
+
+// ── Phase 6.2 — Texture atlas viewer ────────────────────────────────────────
+//
+// A read-only GPU-perf observability tab that answers "is my SDF atlas
+// thrashing?". It renders the render layer's texture-atlas inventory —
+// per-atlas dimensions, page count, live entry count, and a shelf-packer
+// occupancy bar. Like the Phase 6.1 pass viewer and Phase 5.2
+// reconciliation tab it takes over the property-panel region; it has no
+// interactive controls (purely informational), so there are no hit-rects
+// to record. Degrades gracefully to a "GPU atlas unavailable" line when
+// no inventory is wired (headless / GPU-off builds).
+
+void InspectorOverlay::paint_atlas_tab(Canvas& canvas, float x, float y,
+                                       float w, float h) {
+    // Occupancy bar colors: green when there's headroom, amber when the
+    // atlas is filling, red when nearly full (thrash risk). The amber/red
+    // pair matches the drift drawer + reconciliation tab so the GPU-perf
+    // surfaces read consistently.
+    const Color kBarTrack = Color::rgba(0.20f, 0.20f, 0.24f, 1.0f);
+    const Color kBarLow   = Color::rgba(0.35f, 0.85f, 0.45f, 1.0f);
+    const Color kBarMid   = Color::rgba(0.95f, 0.65f, 0.25f, 1.0f);
+    const Color kBarHigh  = Color::rgba(0.95f, 0.40f, 0.38f, 1.0f);
+
+    canvas.set_font("monospace", kFontSize);
+
+    // Section heading.
+    canvas.set_fill_color(kHighlightStroke);
+    canvas.fill_text("Texture Atlases (A)", x, y + 11);
+
+    // No inventory wired — graceful empty state. This is the headless /
+    // GPU-off path; the tab must never crash here.
+    if (!atlas_inventory_ || atlas_inventory_->empty()) {
+        atlas_row_count_ = 0;
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text("GPU atlas unavailable", x, y + 11 + kRowHeight);
+        return;
+    }
+
+    const auto& atlases = atlas_inventory_->atlases();
+    atlas_row_count_ = atlases.size();
+
+    // Summary line: total pages + total packed entries, right-aligned in
+    // the header — mirrors the reconciliation tab's per-status summary.
+    {
+        std::ostringstream summary;
+        summary << atlas_inventory_->total_pages() << " pages  "
+                << atlas_inventory_->total_entries() << " entries";
+        canvas.set_fill_color(kPanelDim);
+        float sw = canvas.measure_text(summary.str());
+        canvas.fill_text(summary.str(), x + w - sw, y + 11);
+    }
+
+    // Each atlas occupies a two-line row: a label/dimensions/entries
+    // line, then an occupancy bar with a percentage readout.
+    constexpr float kAtlasRowH = kRowHeight * 2.0f + 6.0f;
+
+    canvas.save();
+    float list_top = y + kRowHeight;
+    canvas.clip_rect(x, list_top, w, h - kRowHeight);
+
+    float row_y = list_top - atlas_scroll_y_;
+    for (const auto& a : atlases) {
+        const bool visible = row_y > list_top - kAtlasRowH && row_y < y + h;
+        if (visible) {
+            // Line 1: "<label>  <W>x<H>" left, "<N> entries" right.
+            std::ostringstream dims;
+            dims << a.label << "  " << a.width << "x" << a.height;
+            if (a.pages > 1) dims << " x" << a.pages << "p";
+            canvas.set_fill_color(kPanelText);
+            canvas.fill_text(dims.str(), x, row_y + 12);
+
+            std::ostringstream ent;
+            ent << a.entries << " entries";
+            canvas.set_fill_color(kPanelDim);
+            float ew = canvas.measure_text(ent.str());
+            canvas.fill_text(ent.str(), x + w - ew, row_y + 12);
+
+            // Line 2: occupancy bar + percent readout.
+            const int pct = a.occupancy_percent();
+            const Color& fill = pct >= 85 ? kBarHigh
+                              : pct >= 60 ? kBarMid
+                                          : kBarLow;
+            const float bar_y = row_y + kRowHeight;
+            const float bar_h = 8.0f;
+            std::ostringstream pctstr;
+            pctstr << pct << "%";
+            const float pct_w = canvas.measure_text(pctstr.str()) + 6.0f;
+            const float bar_w = w - pct_w;
+
+            canvas.set_fill_color(kBarTrack);
+            canvas.fill_rounded_rect(x, bar_y, bar_w, bar_h, 2.0f);
+            float frac = static_cast<float>(pct) / 100.0f;
+            if (frac > 0.0f) {
+                canvas.set_fill_color(fill);
+                canvas.fill_rounded_rect(x, bar_y,
+                                         std::max(2.0f, bar_w * frac),
+                                         bar_h, 2.0f);
+            }
+            canvas.set_fill_color(fill);
+            canvas.fill_text(pctstr.str(), x + w - pct_w + 6.0f,
+                             bar_y + bar_h);
+        }
+        row_y += kAtlasRowH;
+    }
+
+    canvas.restore();
+}
+
+// ── Phase 2 — Drift drawer ──────────────────────────────────────────────────
+//
+// A collapsible warning panel that lists tweaks whose anchor / property
+// no longer maps to the live design. Header is always shown when drift
+// exists (so the count badge is visible); the body — one row per
+// orphaned/drifted tweak — only renders when expanded. Each row shows
+// the anchor, the dotted property path, the stored value, and a reason
+// tag. Clicking the header chevron toggles the drawer.
+
+float InspectorOverlay::paint_drift_drawer(Canvas& canvas, float x, float y,
+                                           float w) {
+    drift_header_hit_ = {};  // reset; repopulated below if we paint.
+    if (drifted_.empty()) return 0.0f;
+
+    const Color kDriftBg     = Color::rgba(0.22f, 0.06f, 0.07f, 0.95f);
+    const Color kDriftBorder = Color::rgba(0.95f, 0.32f, 0.30f, 0.85f);
+    const Color kDriftText   = Color::rgba(0.98f, 0.72f, 0.70f, 1.0f);
+    const Color kDriftReason = Color::rgba(0.95f, 0.55f, 0.30f, 1.0f);
+
+    constexpr float kHeaderH = 22.0f;
+    constexpr float kDriftRowH = 30.0f;
+    // Cap the body so a large drift list never eats the whole panel.
+    const std::size_t kMaxRows = 6;
+    const std::size_t shown_rows =
+        drift_drawer_open_ ? std::min(drifted_.size(), kMaxRows) : 0;
+    const bool truncated = drifted_.size() > kMaxRows;
+    float body_h = static_cast<float>(shown_rows) * kDriftRowH;
+    if (drift_drawer_open_ && truncated) body_h += 16.0f;  // "+N more" line
+    float total_h = kHeaderH + body_h + 4.0f;
+
+    canvas.set_font("monospace", kFontSize);
+
+    // ── Header ──────────────────────────────────────────────────────
+    canvas.set_fill_color(kDriftBg);
+    canvas.fill_rect(x, y, w, kHeaderH);
+    canvas.set_stroke_color(kDriftBorder);
+    canvas.set_line_width(1.0f);
+    canvas.stroke_rect(x, y, w, kHeaderH);
+
+    // The whole header is the toggle target (generous hit box).
+    drift_header_hit_ = {x, y, w, kHeaderH};
+
+    auto chevron = drift_drawer_open_ ? "\xe2\x96\xbc" : "\xe2\x96\xb6";
+    canvas.set_fill_color(kDriftBorder);
+    canvas.fill_text(chevron, x + 4, y + 15);
+
+    std::string title = "\xe2\x9a\xa0 Drift — " +
+                        std::to_string(drifted_.size()) +
+                        (drifted_.size() == 1 ? " orphaned tweak"
+                                              : " orphaned tweaks");
+    canvas.set_fill_color(kDriftText);
+    canvas.fill_text(title, x + 18, y + 15);
+
+    if (!drift_drawer_open_) return total_h;
+
+    // ── Body — one row per drifted/orphaned tweak ──────────────────
+    canvas.set_fill_color(Color::rgba(0.14f, 0.05f, 0.06f, 0.95f));
+    canvas.fill_rect(x, y + kHeaderH, w, body_h);
+
+    float row_y = y + kHeaderH;
+    for (std::size_t i = 0; i < shown_rows; ++i) {
+        const auto& d = drifted_[i];
+
+        // Left red marker stripe — matches Phase 2.5's planned
+        // drift-row styling so the two panels read consistently.
+        canvas.set_fill_color(kDriftBorder);
+        canvas.fill_rect(x, row_y + 2, 2.0f, kDriftRowH - 4);
+
+        // Line 1: anchor + reason tag.
+        std::string anchor = d.anchor_id;
+        if (anchor.size() > 28) anchor = anchor.substr(0, 27) + "\xe2\x80\xa6";
+        canvas.set_fill_color(kDriftText);
+        canvas.fill_text(anchor, x + 8, row_y + 13);
+
+        std::string reason = TweakStore::drift_reason_str(d.reason);
+        canvas.set_fill_color(kDriftReason);
+        float rw = canvas.measure_text(reason);
+        canvas.fill_text(reason, x + w - rw - 4, row_y + 13);
+
+        // Line 2: property path = stored value.
+        std::string value_str;
+        if (d.value.isString()) {
+            value_str = std::string(d.value.getString());
+        } else {
+            try {
+                value_str = choc::json::toString(d.value);
+            } catch (...) {
+                value_str = "?";
+            }
+        }
+        std::string detail = d.property_path + " = " + value_str;
+        if (detail.size() > 40) detail = detail.substr(0, 39) + "\xe2\x80\xa6";
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text(detail, x + 8, row_y + 25);
+
+        row_y += kDriftRowH;
+    }
+
+    if (truncated) {
+        canvas.set_fill_color(kPanelDim);
+        canvas.fill_text("\xe2\x80\xa6 +" +
+                             std::to_string(drifted_.size() - kMaxRows) +
+                             " more (see `pulp tweaks diff`)",
+                         x + 8, row_y + 12);
+    }
+
+    return total_h;
+}
+
+}  // namespace pulp::inspect
